+++ argos/widgets/preferences.py
@@ -8,18 +8,11 @@ from gi.repository import Gio, GLib, GObject, Gtk
 if TYPE_CHECKING:
     from argos.app import Application
 
-from argos.widgets.utils import ALBUM_SORT_CHOICES
-
 LOGGER = logging.getLogger(__name__)
 
 _ = gettext.gettext
 
 SECONDS_PER_DAY = 3600 * 24
-
-
-class AlbumSortStoreColumns(IntEnum):
-    TEXT = 0
-    ID = 1
 
 
 @Gtk.Template(resource_path="/io/github/orontee/Argos/ui/preferences.ui")
@@ -36,7 +29,6 @@ class PreferencesWindow(Gtk.Window):
     mopidy_jellyfin_label: Gtk.Label = Gtk.Template.Child()
     mopidy_podcast_switch: Gtk.Switch = Gtk.Template.Child()
     mopidy_podcast_label: Gtk.Label = Gtk.Template.Child()
-    album_sort_view: Gtk.TreeView = Gtk.Template.Child()
     history_playlist_check_button: Gtk.CheckButton = Gtk.Template.Child()
     history_playlist_max_length_label: Gtk.Label = Gtk.Template.Child()
     history_playlist_max_length_button: Gtk.SpinButton = Gtk.Template.Child()
@@ -54,17 +46,6 @@ class PreferencesWindow(Gtk.Window):
         self.set_wmclass("Argos", "preferences")
         self._model = application.model
         self._albums_image_size_scale_jumped_id: Optional[int] = None
-
-        store = Gtk.ListStore(str, str)
-        self.album_sort_view.set_model(store)
-        self.album_sort_view.set_activate_on_single_click(True)
-        self.album_sort_view.set_headers_visible(False)
-        renderer = Gtk.CellRendererText()
-        column = Gtk.TreeViewColumn(
-            cell_renderer=renderer, text=AlbumSortStoreColumns.TEXT
-        )
-        self.album_sort_view.append_column(column)
-        self.album_sort_view.set_enable_search(False)
 
         self.connection_warning_label.set_visible(
             self._model.network_available and not self._model.connected
@@ -86,17 +67,6 @@ class PreferencesWindow(Gtk.Window):
 
         mopidy_podcast = self._settings.get_boolean("mopidy-podcast")
         self.mopidy_podcast_switch.set_active(mopidy_podcast)
-
-        album_sort_id = self._settings.get_string("album-sort")
-        to_select: Optional[Gtk.TreeIter] = None
-        for id, name in ALBUM_SORT_CHOICES.items():
-            store_iter = store.append([name, id])
-            if id == album_sort_id:
-                to_select = store_iter
-
-        if to_select is not None:
-            selection = self.album_sort_view.get_selection()
-            selection.select_iter(to_select)
 
         history_playlist = self._settings.get_boolean("history-playlist")
         self.history_playlist_check_button.set_active(history_playlist)
@@ -193,9 +163,6 @@ class PreferencesWindow(Gtk.Window):
         self.mopidy_podcast_switch.connect(
             "notify::active", self.on_mopidy_podcast_switch_activated
         )
-        self.album_sort_view.connect(
-            "row-activated", self.on_album_sort_view_row_activated
-        )
         self.history_playlist_check_button.connect(
             "toggled", self.on_history_playlist_check_button_toggled
         )
@@ -288,14 +255,6 @@ class PreferencesWindow(Gtk.Window):
     ) -> None:
         mopidy_podcast = switch.get_active()
         self._settings.set_boolean("mopidy-podcast", mopidy_podcast)
-
-    def on_album_sort_view_row_activated(
-        self, _1: Gtk.TreeView, _2: Gtk.TreePath, _3: Gtk.TreeViewColumn
-    ) -> None:
-        selection = self.album_sort_view.get_selection()
-        store, store_iter = selection.get_selected()
-        album_sort_id = store.get_value(store_iter, AlbumSortStoreColumns.ID)
-        self._settings.set_string("album-sort", album_sort_id)
 
     def on_history_playlist_check_button_toggled(self, button: Gtk.CheckButton) -> None:
         history_playlist = button.get_active()